Fixed implementation for using (s3, bunnycdn)
For an implementation in Laravel for Bunny CDN, i also had this issue with Amazon S3 the image optimizer could not write to the bucket and to the local file system because the file already exists. To make this wordt i replaced the writeStream with putStream. because it describes "Write or Update Files using a stream"